+++ Sources/Sentry/SentryOptions.m
@@ -118,9 +118,10 @@ SentryOptions ()
 - (void)validateOptions:(NSDictionary<NSString *, id> *)options
        didFailWithError:(NSError *_Nullable *_Nullable)error
 {
-    if (nil != options[@"debug"]) {
-        self.debug = [options[@"debug"] boolValue];
-    }
+    NSPredicate *isNSString = [NSPredicate predicateWithBlock:^BOOL(
+        id object, NSDictionary *bindings) { return [object isKindOfClass:[NSString class]]; }];
+
+    [self setBool:options[@"debug"] block:^(BOOL value) { self->_debug = value; }];
 
     if ([options[@"diagnosticLevel"] isKindOfClass:[NSString class]]) {
         for (SentryLevel level = 0; level <= kSentryLevelFatal; level++) {
@@ -132,9 +133,8 @@ SentryOptions ()
     }
 
     NSString *dsn = @"";
-    if (nil != [options valueForKey:@"dsn"] &&
-        [[options valueForKey:@"dsn"] isKindOfClass:[NSString class]]) {
-        dsn = [options valueForKey:@"dsn"];
+    if (nil != options[@"dsn"] && [options[@"dsn"] isKindOfClass:[NSString class]]) {
+        dsn = options[@"dsn"];
     }
 
     self.parsedDsn = [[SentryDsn alloc] initWithString:dsn didFailWithError:error];
@@ -151,9 +151,7 @@ SentryOptions ()
         self.dist = options[@"dist"];
     }
 
-    if (nil != options[@"enabled"]) {
-        self.enabled = [options[@"enabled"] boolValue];
-    }
+    [self setBool:options[@"enabled"] block:^(BOOL value) { self->_enabled = value; }];
 
     if ([options[@"maxBreadcrumbs"] isKindOfClass:[NSNumber class]]) {
         self.maxBreadcrumbs = [options[@"maxBreadcrumbs"] unsignedIntValue];
@@ -163,75 +161,63 @@ SentryOptions ()
         self.maxCacheItems = [options[@"maxCacheItems"] unsignedIntValue];
     }
 
-    if (nil != options[@"beforeSend"]) {
+    if ([self isBlock:options[@"beforeSend"]]) {
         self.beforeSend = options[@"beforeSend"];
     }
 
-    if (nil != options[@"beforeBreadcrumb"]) {
+    if ([self isBlock:options[@"beforeBreadcrumb"]]) {
         self.beforeBreadcrumb = options[@"beforeBreadcrumb"];
     }
 
-    if (nil != options[@"onCrashedLastRun"]) {
+    if ([self isBlock:options[@"onCrashedLastRun"]]) {
         self.onCrashedLastRun = options[@"onCrashedLastRun"];
     }
 
-    if (nil != options[@"integrations"]) {
-        self.integrations = options[@"integrations"];
+    if ([options[@"integrations"] isKindOfClass:[NSArray class]]) {
+        self.integrations = [options[@"integrations"] filteredArrayUsingPredicate:isNSString];
     }
 
-    NSNumber *sampleRate = options[@"sampleRate"];
-    if (nil != sampleRate) {
-        self.sampleRate = sampleRate;
+    if ([options[@"sampleRate"] isKindOfClass:[NSNumber class]]) {
+        self.sampleRate = options[@"sampleRate"];
     }
 
-    if (nil != options[@"enableAutoSessionTracking"]) {
-        self.enableAutoSessionTracking = [options[@"enableAutoSessionTracking"] boolValue];
-    }
+    [self setBool:options[@"enableAutoSessionTracking"]
+            block:^(BOOL value) { self->_enableAutoSessionTracking = value; }];
 
-    if (nil != options[@"enableOutOfMemoryTracking"]) {
-        self.enableOutOfMemoryTracking = [options[@"enableOutOfMemoryTracking"] boolValue];
-    }
+    [self setBool:options[@"enableOutOfMemoryTracking"]
+            block:^(BOOL value) { self->_enableOutOfMemoryTracking = value; }];
 
-    if (nil != options[@"sessionTrackingIntervalMillis"]) {
+    if ([options[@"sessionTrackingIntervalMillis"] isKindOfClass:[NSNumber class]]) {
         self.sessionTrackingIntervalMillis =
             [options[@"sessionTrackingIntervalMillis"] unsignedIntValue];
     }
 
-    if (nil != options[@"attachStacktrace"]) {
-        self.attachStacktrace = [options[@"attachStacktrace"] boolValue];
-    }
+    [self setBool:options[@"attachStacktrace"]
+            block:^(BOOL value) { self->_attachStacktrace = value; }];
 
-    if (nil != options[@"stitchAsyncCode"]) {
-        self.stitchAsyncCode = [options[@"stitchAsyncCode"] boolValue];
-    }
+    [self setBool:options[@"stitchAsyncCode"]
+            block:^(BOOL value) { self->_stitchAsyncCode = value; }];
 
-    if (nil != options[@"maxAttachmentSize"]) {
+    if ([options[@"maxAttachmentSize"] isKindOfClass:[NSNumber class]]) {
         self.maxAttachmentSize = [options[@"maxAttachmentSize"] unsignedIntValue];
     }
 
-    if (nil != options[@"sendDefaultPii"]) {
-        self.sendDefaultPii = [options[@"sendDefaultPii"] boolValue];
+    [self setBool:options[@"sendDefaultPii"]
+            block:^(BOOL value) { self->_sendDefaultPii = value; }];
+
+    [self setBool:options[@"enableAutoPerformanceTracking"]
+            block:^(BOOL value) { self->_enableAutoPerformanceTracking = value; }];
+
+    [self setBool:options[@"enableNetworkTracking"]
+            block:^(BOOL value) { self->_enableNetworkTracking = value; }];
+
+    if ([options[@"tracesSampleRate"] isKindOfClass:[NSNumber class]]) {
+        self.tracesSampleRate = options[@"tracesSampleRate"];
     }
 
-    if (nil != options[@"enableAutoPerformanceTracking"]) {
-        self.enableAutoPerformanceTracking = [options[@"enableAutoPerformanceTracking"] boolValue];
-    }
-
-    if (nil != options[@"enableNetworkTracking"]) {
-        self.enableNetworkTracking = [options[@"enableNetworkTracking"] boolValue];
-    }
-
-    NSNumber *tracesSampleRate = options[@"tracesSampleRate"];
-    if (nil != tracesSampleRate) {
-        self.tracesSampleRate = tracesSampleRate;
-    }
-
-    if (nil != options[@"tracesSampler"]) {
+    if ([self isBlock:options[@"tracesSampler"]]) {
         self.tracesSampler = options[@"tracesSampler"];
     }
-
-    NSPredicate *isNSString = [NSPredicate predicateWithBlock:^BOOL(
-        id object, NSDictionary *bindings) { return [object isKindOfClass:[NSString class]]; }];
 
     if ([options[@"inAppIncludes"] isKindOfClass:[NSArray class]]) {
         NSArray<NSString *> *inAppIncludes =
@@ -247,12 +233,18 @@ SentryOptions ()
         self.urlSessionDelegate = options[@"urlSessionDelegate"];
     }
 
-    if (options[@"experimentalEnableTraceSampling"] != nil) {
-        _experimentalEnableTraceSampling = [options[@"experimentalEnableTraceSampling"] boolValue];
-    }
+    [self setBool:options[@"experimentalEnableTraceSampling"]
+            block:^(BOOL value) { self->_experimentalEnableTraceSampling = value; }];
 
-    if (options[@"enableSwizzling"] != nil) {
-        _enableSwizzling = [options[@"enableSwizzling"] boolValue];
+    [self setBool:options[@"enableSwizzling"]
+            block:^(BOOL value) { self->_enableSwizzling = value; }];
+}
+
+- (void)setBool:(id)value block:(void (^)(BOOL))block
+{
+    // Entries in the dictionary can be NSNull. Especially, on React-Native, this can happen.
+    if (value != nil && ![value isEqual:[NSNull null]]) {
+        block([value boolValue]);
     }
 }
 
@@ -304,6 +296,28 @@ SentryOptions ()
 {
     return (_tracesSampleRate != nil && [_tracesSampleRate doubleValue] > 0)
         || _tracesSampler != nil;
+}
+
+/**
+ * Checks if the passed in block is actually of type block. We can't check if the block matches a
+ * specific block without some complex objc runtime method calls and therefore we only check if its
+ * a block or not. Assigning a wrong block to the SentryOption blocks still could lead to crashes at
+ * runtime, but when someone uses the initWithDict they should better know what they are doing.
+ *
+ * Taken from https://gist.github.com/steipete/6ee378bd7d87f276f6e0
+ */
+- (BOOL)isBlock:(nullable id)block
+{
+    static Class blockClass;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        blockClass = [^{} class];
+        while ([blockClass superclass] != NSObject.class) {
+            blockClass = [blockClass superclass];
+        }
+    });
+
+    return [block isKindOfClass:blockClass];
 }
 
 @end